feature(today-avarage-component): removing run failed
