+++ challenge-catar Watch App/Views/Components/TextView/TextType.swift
@@ -5,7 +5,7 @@
 //  Created by alexdamascena on 27/11/22.
 //
 
-import Foundation
+import SwiftUI
 
 enum TextType {
